refactor: renames eval module for clarity
And to avoid conflict with `rlplg`
