+++ src/daaf/task.py
@@ -9,9 +9,8 @@ from typing import Any, Callable, Generator, Mapping, Optional, Sequence, Tuple
 import gymnasium as gym
 from rlplg import core, envplay, envsuite
 from rlplg.learning import utils
-from rlplg.learning.tabular import policies
 
-from daaf import constants, expconfig, options, replay_mapper
+from daaf import constants, replay_mapper
 
 
 def create_env_spec(
@@ -104,25 +103,6 @@ def create_trajectory_mappers(
             Choices: {constants.AGGREGATE_MAPPER_METHODS}"""
         )
     return mappers
-
-
-def create_eval_policy(
-    env_spec: core.EnvSpec, daaf_config: expconfig.DaafConfig
-) -> core.PyPolicy:
-    """
-    Creates a policy to be evaluated.
-    """
-    if daaf_config.policy_type == constants.OPTIONS_POLICY:
-        return options.UniformlyRandomCompositeActionPolicy(
-            actions=tuple(range(env_spec.mdp.env_desc.num_actions)),
-            options_duration=daaf_config.reward_period,
-        )
-    elif daaf_config.policy_type == constants.SINGLE_STEP_POLICY:
-        return policies.PyRandomPolicy(
-            num_actions=env_spec.mdp.env_desc.num_actions,
-            emit_log_probability=True,
-        )
-    raise ValueError(f"Unknown policy {daaf_config.policy_type}")
 
 
 def create_generate_episodes_fn(
